fix ui pepe
