+++ src/pages/Pool-V3/components/PoolList/index.tsx
@@ -145,19 +145,6 @@ const PoolList = ({ search }) => {
         };
       })
       .sort((a, b) => {
-        if (isShowPepeOnTop) {
-          const PepePoolKey = `${PEPE_ORAICHAIN_EXT_DENOM}-${USDC_CONTRACT}-3000000000-100`;
-          // const PepePoolKey =
-          //   'factory/orai1wuvhex9xqs3r539mvc6mtm7n20fcj3qr2m0y9khx6n5vtlngfzes3k0rq9/extPEPE-orai15un8msx3n5zf9ahlxmfeqd2kwa5wm0nrpxer304m9nd5q6qq0g6sku5pdd-3000000000-100';
-          // FIXME: push pepe/usdc on top
-          if (a.poolKey === PepePoolKey && b.poolKey !== PepePoolKey) {
-            return -1; // Push PepePoolKey elements to the top
-          }
-          if (a.poolKey !== PepePoolKey && b.poolKey === PepePoolKey) {
-            return 1; // Keep non-'a' elements below 'a'
-          }
-        }
-
         switch (sortField) {
           case PoolColumnHeader.LIQUIDITY:
             return (
