+++ src/store/constants.ts
@@ -1,5 +1,5 @@
 // change version when update, add state redux-persist storage
-export const PERSIST_VER = 2.3;
+export const PERSIST_VER = 2.4;
 
 export const PERSIST_CONFIG_KEY = 'root';
 export const ADDRESS_BOOK_KEY_BACKUP = 'addressBookBackup';
